refactor(utils): move download helpers into module
Refs python-poetry/poetry#11004 review feedback.
