+++ src/poetry/utils/helpers.py
@@ -9,18 +9,15 @@ import stat
 import sys
 import tarfile
 import tempfile
+import warnings as _warnings
 import zipfile
 
 from collections.abc import Mapping
 from contextlib import contextmanager
-from contextlib import suppress
-from functools import cached_property
 from pathlib import Path
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import overload
-
-from poetry.utils.constants import REQUESTS_TIMEOUT
 
 
 if TYPE_CHECKING:
@@ -30,11 +27,6 @@ if TYPE_CHECKING:
     from types import TracebackType
 
     from poetry.core.packages.package import Package
-    from requests import Response
-    from requests import Session
-
-    from poetry.utils.authenticator import Authenticator
-
 logger = logging.getLogger(__name__)
 prioritised_hash_types: tuple[str, ...] = tuple(
     t
@@ -122,125 +114,6 @@ def merge_dicts(d1: dict[str, Any], d2: dict[str, Any]) -> None:
 
 class HTTPRangeRequestSupportedError(Exception):
     """Raised when server unexpectedly supports byte ranges."""
-
-
-def download_file(
-    url: str,
-    dest: Path,
-    *,
-    session: Authenticator | Session | None = None,
-    chunk_size: int = 1024,
-    raise_accepts_ranges: bool = False,
-    max_retries: int = 0,
-) -> None:
-    from poetry.puzzle.provider import Indicator
-
-    downloader = Downloader(url, dest, session, max_retries=max_retries)
-
-    if raise_accepts_ranges and downloader.accepts_ranges:
-        raise HTTPRangeRequestSupportedError(f"URL {url} supports range requests.")
-
-    set_indicator = False
-    with Indicator.context() as update_context:
-        update_context(f"Downloading {url}")
-
-        total_size = downloader.total_size
-        if total_size > 0:
-            fetched_size = 0
-            last_percent = 0
-
-            # if less than 1MB, we simply show that we're downloading
-            # but skip the updating
-            set_indicator = total_size > 1024 * 1024
-
-        for fetched_size in downloader.download_with_progress(chunk_size):
-            if set_indicator:
-                percent = (fetched_size * 100) // total_size
-                if percent > last_percent:
-                    last_percent = percent
-                    update_context(f"Downloading {url} {percent:3}%")
-
-
-class Downloader:
-    def __init__(
-        self,
-        url: str,
-        dest: Path,
-        session: Authenticator | Session | None = None,
-        max_retries: int = 0,
-    ):
-        self._dest = dest
-        self._max_retries = max_retries
-        if session is None:
-            from poetry.utils.authenticator import get_default_authenticator
-
-            session = get_default_authenticator()
-        self._session = session
-        self._url = url
-        self._response = self._get()
-
-    @cached_property
-    def accepts_ranges(self) -> bool:
-        return self._response.headers.get("Accept-Ranges") == "bytes"
-
-    @cached_property
-    def total_size(self) -> int:
-        total_size = 0
-        if "Content-Length" in self._response.headers:
-            with suppress(ValueError):
-                total_size = int(self._response.headers["Content-Length"])
-        return total_size
-
-    def _get(self, start: int = 0) -> Response:
-        headers = {"Accept-Encoding": "Identity"}
-        if start > 0:
-            headers["Range"] = f"bytes={start}-"
-
-        response = self._session.get(
-            self._url, stream=True, headers=headers, timeout=REQUESTS_TIMEOUT
-        )
-        try:
-            response.raise_for_status()
-            return response
-        except BaseException:
-            response.close()
-            raise
-
-    def _iter_content_with_resume(self, chunk_size: int) -> Iterator[bytes]:
-        fetched_size = 0
-        retries = 0
-        while True:
-            try:
-                with self._response:
-                    for chunk in self._response.iter_content(chunk_size=chunk_size):
-                        yield chunk
-                        fetched_size += len(chunk)
-            except _resumable_errors():
-                if (
-                    retries < self._max_retries
-                    and self.accepts_ranges
-                    and fetched_size > 0
-                ):
-                    # only retry if server supports byte ranges
-                    # and we have fetched at least one chunk
-                    # otherwise, we should just fail
-                    retries += 1
-                    self._response = self._get(fetched_size)
-                    continue
-                raise
-            else:
-                break
-
-    def download_with_progress(self, chunk_size: int = 1024) -> Iterator[int]:
-        fetched_size = 0
-        from requests.utils import atomic_open
-
-        with atomic_open(self._dest) as f:
-            for chunk in self._iter_content_with_resume(chunk_size=chunk_size):
-                if chunk:
-                    f.write(chunk)
-                    fetched_size += len(chunk)
-                    yield fetched_size
 
 
 def get_package_version_display_string(
@@ -460,34 +333,22 @@ def extractall(source: Path, dest: Path, zip: bool) -> None:
                 archive.extractall(dest, members=safe_members)
 
 
-def _resumable_errors() -> tuple[type[BaseException], ...]:
-    from requests.exceptions import ChunkedEncodingError
-    from requests.exceptions import ConnectionError as RequestsConnectionError
-
-    return (ChunkedEncodingError, RequestsConnectionError)
-
-
-_LAZY_REEXPORTS = {
-    "ChunkedEncodingError": ("requests.exceptions", "ChunkedEncodingError"),
-    "ConnectionError": ("requests.exceptions", "ConnectionError"),
-    "atomic_open": ("requests.utils", "atomic_open"),
-    "get_default_authenticator": (
-        "poetry.utils.authenticator",
-        "get_default_authenticator",
-    ),
-}
+_DEPRECATED_DOWNLOAD_EXPORTS = {"Downloader", "download_file"}
 
 
 def __getattr__(name: str) -> Any:
-    target = _LAZY_REEXPORTS.get(name)
-    if target is None:
-        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-    from importlib import import_module
+    if name in _DEPRECATED_DOWNLOAD_EXPORTS:
+        _warnings.warn(
+            f"Importing `{name}` from `poetry.utils.helpers` is deprecated; "
+            f"use `poetry.utils.download.{name}` instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from poetry.utils import download
 
-    value = getattr(import_module(target[0]), target[1])
-    globals()[name] = value
-    return value
+        return getattr(download, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 def __dir__() -> list[str]:
-    return sorted([*globals(), *_LAZY_REEXPORTS])
+    return sorted([*globals(), *_DEPRECATED_DOWNLOAD_EXPORTS])
